Remove xmpp.patch (#15149, #15162)
xmpp.patch was doing two things: setting the default server to
jabber.ccc.de, and setting the XMPP resource to an empty string. #15149
lists the reasons why we shouldn't use jabber.ccc.de as the default
server, and since we are no longer pretending not to be Instantbird, we
should not clear the resource (#15162).